+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Warning of 398 day limit for certificates issued after 2020/9/1
 * Added environment variable for amount of attempts for ssl renegotiation check
 * Added --user-agent argument to support using a custom User Agent
+* CSV and JSON output: Remove spaces and PEM header from certificiates
 
 ### Features implemented / improvements in 3.0